Use SQL-92 and loosely apply style guide
As a step towards being DB agnostic, the schema file is written in
standard SQL-92. It is currently only tested with SQLite so milage may
still vary.

The schema file also loosely follows https://www.sqlstyle.guide/ now.
One rule is purposefully not adhered to: the primary key declaration
is not the first thing after the CREATE TABLE. That ordering is
explicitly not supported by SQLite.

This also introduces a SQL linting tool for development.